feat: challenge 6 ts local backend
